requirements.txt
beautifulsoup4==4.12.2
boto3==1.34.14
botocore==1.34.14
gunicorn==20.1.0
python-dotenv
Flask==3.0.0
folium==0.15.1
MarkupSafe==2.1.3
pandas==2.1.4
requests==2.31.0